Form factor of nanodisc with endcaps works. Intensity wrong for now

+++ include/Bead.h
@@ -8,16 +8,16 @@ class Bead {
     Bead();
     ~Bead();
 
-    float x; /* position along x */
-    float y; /* position along y */
-    float z; /* position along z */
-    float rho; /* excess scattering length */
-    float v; /* volume */
-    float nn; /* number of neighbors */
+    double x; /* position along x */
+    double y; /* position along y */
+    double z; /* position along z */
+    double rho; /* excess scattering length */
+    double v; /* volume */
+    double nn; /* number of neighbors */
 
     bool position_assigned;
 
-    void assign_position( float, float, float );
+    void assign_position( double, double, double );
     void assign_volume_and_scattlen( const std::string& );
 
 };
